Use log in workspace root
The --log-base argument is handled differently than other base-folder arguments.
Its default value is None, meaning that the environment variable COLCON_LOG_PATH
or the default path is used if the command-line argument is missing.
As we cannot easily change the default path (which is hardcoded in colcon-core),
we override the environment variable instead.